fix home page include material app.
