+++ lib/main.dart
@@ -2,26 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_demo/home/home_page.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'counter/counter_page.dart';
-
 void main() {
-  runApp(const ProviderScope(child: MyApp()));
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      routes: {
-        '/': (context) => const HomePage(),
-        CounterPage.route: (context) => const CounterPage(),
-      },
-    );
-  }
+  runApp(const ProviderScope(child: HomePage()));
 }
